requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pymongo==3.12.0
python-dotenv==0.19.0
python-jose
urllib3
passlib==1.7.4
bcrypt==3.2.0
pyrebase4
setuptools

